Move prusalink base entity to separate module
