+++ homeassistant/components/prusalink/camera.py
@@ -9,9 +9,9 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import PrusaLinkEntity
 from .const import DOMAIN
 from .coordinator import JobUpdateCoordinator
+from .entity import PrusaLinkEntity
 
 
 async def async_setup_entry(
